+++ docs/source/change_log.rst
@@ -16,6 +16,7 @@ Version 5.0.0
 * **Breaking:** Changed the precedence of operators to be more aligned with common programming languages
 * **Breaking:** The :ref:`debug-repl` utility now requires development dependencies to be installed
 * **Breaking:** BYTES literals now throw exceptions for invalid escape sequences
+* Added support for raw STRING literals using the ``r`` prefix (e.g. ``r'\w+'``)
 
 Version 4.x.x
 -------------